FoodController and change database

+++ src/main/java/cn/edu/uestc/rms/model/Revenue.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Revenue {
-    private int revenueId;
-    private int orderId;
-    private double money;
+    private Integer revenueId;
+    private Integer orderId;
+    private Integer money;
     private String time;
 }
